requirements.txt
anyio==4.4.0
asgiref==3.8.1
Authlib==1.3.1
azure-core==1.30.2
azure-storage-blob==12.20.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
Django==5.0.6
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-filter==24.2
django-rest-passwordreset==1.4.1
django-storages==1.14.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.3.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
